pyproject.toml
    [project]
name = "online-retail-platform"
version = "0.1.0"
description = ""
authors = [
    { name = "yakhovetso" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-unfold (>=0.64.0,<0.65.0)",
    "djoser (>=2.3.3,<3.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "qrcode (>=8.2,<9.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "django-environ (>=0.12.0,<0.13.0)"
]

[tool.poetry.group.dev.dependencies]
faker = "^37.5.3"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
line_length = 120
lines_after_imports = 2
lines_between_sections = 1
lines_between_types = 0
force_alphabetical_sort_within_sections = true
ensure_newline_before_comments = true
combine_as_imports = true
combine_star = true
skip_gitignore = true
balanced_wrapping = true
force_grid_wrap = 2
src_paths = ["core", ]
skip_glob = [
    ".git",
    ".idea",
    "venv",
    ".venv",
]
known_django = ["django", "rest_framework"]
known_first_party = ["core"]
sections = [
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
